requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Django==4.2.6
django-geojson==4.0.0
django-leaflet==0.29.0
docopt==0.6.2
fastjsonschema==2.19.1
frozenlist==1.4.1
GDAL==3.6.2
idna==3.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.1
multidict==6.0.5
nbformat==5.9.2
numpy
pigar==2.1.2
Pillow
platformdirs==4.2.0
psycopg==3.1.12
psycopg2-binary==2.9.9
pywin32==306
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
sqlparse==0.4.4
traitlets==5.14.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
yarg==0.1.9
yarl==1.9.4
